done implementing the peparation
