+++ lib/Screen/loan_calculator_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:kodecamp_calculator_app/controllers/calculator_func_controller.dart';
+import 'package:kodecamp_calculator_app/controllers/fd_calculator_controller.dart';
+import 'package:kodecamp_calculator_app/controllers/loan_calculator_controller.dart';
 import 'package:kodecamp_calculator_app/utils/constants/colors.dart';
 import 'package:kodecamp_calculator_app/utils/constants/sizes.dart';
 import 'package:kodecamp_calculator_app/utils/format.dart';
